+++ lua/obsidian/commands/template.lua
@@ -6,8 +6,7 @@ local api = require "obsidian.api"
 return function(_, data)
   local templates_dir = api.templates_dir()
   if not templates_dir then
-    log.err "Templates folder is not defined or does not exist"
-    return
+    return log.err "Templates folder is not defined or does not exist"
   end
 
   -- We need to get this upfront before the picker hijacks the current window.
@@ -35,9 +34,12 @@ return function(_, data)
     return
   end
 
-  picker:find_templates {
+  picker:find_files {
+    prompt_title = "Templates",
     callback = function(path)
       insert_template(path)
     end,
+    dir = templates_dir,
+    no_default_mappings = true,
   }
 end