+++ lua/obsidian/config.lua
@@ -168,8 +168,10 @@ config.default = {
   ---@field create_new? boolean
   completion = (function()
     local has_nvim_cmp, _ = pcall(require, "cmp")
+    local has_blink = pcall(require, "blink.cmp")
     return {
-      nvim_cmp = has_nvim_cmp,
+      nvim_cmp = has_nvim_cmp and not has_blink,
+      blink = has_blink,
       min_chars = 2,
       match_case = true,
       create_new = true,